lit/requirements.txt
absl-py==1.4.0
aqtp==0.0.10
astunparse==1.6.3
attrs==22.2.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
chex==0.1.5
click==8.1.3
cloudpickle==2.2.1
clu==0.0.8
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
dacite==1.8.0
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
einops==0.6.0
etils==1.0.0
exceptiongroup==1.1.0
flatbuffers==23.1.21
flax==0.6.3
flaxformer @ git+https://github.com/google/flaxformer@f7cb09f9a7e2598e76ecd71304b99fd55fb51148
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
jax==0.4.2
jaxlib==0.4.2
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
ml-collections==0.1.1
msgpack==1.0.4
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkg_resources==0.0.0
pluggy==1.0.0
promise==2.3
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.2.0
rsa==4.9
scipy==1.10.0
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.8.2
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-metadata==1.12.0
tensorflow-probability==0.19.0
tensorflow-text==2.11.0
tensorstore==0.1.30
termcolor==2.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
vit-jax @ git+https://github.com/google-research/vision_transformer@a2d1c2c7ee571cae74a36bea55c8d5646774bdd6
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
